requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic[email]
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
schedule==1.2.0

# Face Recognition & Computer Vision
opencv-python==4.8.1.78
deepface==0.0.79
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.1.0
mediapipe==0.10.8
tensorflow-hub==0.15.0
scipy==1.11.4

# QR/Barcode Detection
pyzbar==0.1.9
qrcode[pil]==7.4.2

# Blockchain Integration
fabric-sdk-py==1.0.0

# Notification Services
twilio==8.10.0
python-dotenv==1.0.0

# Additional Utilities
bcrypt==4.1.2
python-dateutil==2.8.2
requests==2.31.0
